Epic Power: Mass Impalement
Very little information in the 3.5 edition source books on epic psionic powers, so I thought I would develop one of my own. This power is fairly straight-forward, dealing massive damage in a large area of effect. While the sheer amount of damage might seem high, in reality this is balanced in the context of a setting with many enemies possessing cold iron or magical weapon-based damage reduction, which is not bypassed by flat 'slashing' or 'piercing' damage regardless of whether or not the source is supernatural.