Astral Construct
This power had to be handled a bit differently than most of the others. Rather than having 1 summon power for each tier like the arcane and divine summoning spells (summon I - IX), I decided that it would be neater to have it be one power that scales (and thus summons a different minion) with each psionic power tier achieved. As a consequence, however, the cost of the power also needs to scale in spite of the fact that mechanically it remains a 1st level power (the meta-script calculating power cost does so by 'spell level' which is stored in the game's 2da framework); thus the power point cost is handled in the power script itself rather than externally in the meta-script.